requirements.txt
pyTelegramBotAPI==4.14.0
gspread==5.12.4
flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-auth==2.25.2
pytz==2023.3.post1
requests==2.28.1
